update account page

+++ routes/admin.php
@@ -6,16 +6,16 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::controller(CountryController::class)->group(function () {
-    Route::get('/admin/countries', 'index')->name('countries.index');
-    Route::post('/admin/countries', 'store')->name('countries.store');
-    Route::put('/admin/countries/{country}', 'update')->name('countries.update');
-    Route::delete('/admin/countries/{country}', 'destroy')->name('countries.destroy');
+    Route::get('/admin/countries', 'index')->name('countries.index')->middleware('auth');
+    Route::post('/admin/countries', 'store')->name('countries.store')->middleware('auth');
+    Route::put('/admin/countries/{country}', 'update')->name('countries.update')->middleware('auth');
+    Route::delete('/admin/countries/{country}', 'destroy')->name('countries.destroy')->middleware('auth');
 });
 
 
 
 Route::controller(SettingController::class)->group(function(){
-    Route::get('/admin/settings', 'index')->name('settings.index');
-    Route::put('/admin/settings/{setting}', 'update')->name('settings.update');
-    Route::delete('/admin/settings/{setting}', 'destroy')->name('settings.destroy');
+    Route::get('/admin/settings', 'index')->name('settings.index')->middleware('auth');
+    Route::put('/admin/settings/{setting}', 'update')->name('settings.update')->middleware('auth');
+    Route::delete('/admin/settings/{setting}', 'destroy')->name('settings.destroy')->middleware('auth');
 });
